make module for writer mock fixtures
